LayoutTests/platform/gtk/fast/canvas/webgl/webgl-depth-texture-expected.txt

This test verifies the functionality of the WEBGL_depth_texture extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
Testing binding enum with extension disabled
PASS gl.texImage2D(gl.TEXTURE_2D, 0, gl.DEPTH_COMPONENT, 1, 1, 0, gl.DEPTH_COMPONENT, gl.UNSIGNED_SHORT, null) generated expected GL error: INVALID_ENUM.
PASS gl.texImage2D(gl.TEXTURE_2D, 0, gl.DEPTH_COMPONENT, 1, 1, 0, gl.DEPTH_COMPONENT, gl.UNSIGNED_INT, null) generated expected GL error: INVALID_ENUM.
PASS No WEBGL_depth_texture support -- this is legal
PASS WEBGL_depth_texture not listed as supported and getExtension failed -- this is legal

PASS successfullyParsed is true

TEST COMPLETE
